aptitude wip

+++ UdpHosts/GameServer/StaticDB/Records/apt/ConditionalBranchCommandDef.cs
@@ -3,6 +3,6 @@ public record class ConditionalBranchCommandDef
 {
     public uint IfChain { get; set; }
     public uint ThenChain { get; set; }
-    public byte ElseChain { get; set; }
+    public uint ElseChain { get; set; }
     public uint Id { get; set; }
 }
